included workaround for missing .tbi file if VCF is empty

+++ Clair3/1.0.4--py39hf5e1c6e_3/VariantCaller.nf
@@ -22,6 +22,12 @@ process VariantCaller {
             --threads=${task.cpus} \
             ${params.optional}
         mv merge_output.vcf.gz ${bam_file.baseName}.vcf.gz
-        mv merge_output.vcf.gz.tbi ${bam_file.baseName}.vcf.gz.tbi
+        if test -f "merge_output.vcf.gz.tbi";
+        then
+            mv merge_output.vcf.gz.tbi ${bam_file.baseName}.vcf.gz.tbi
+        else
+            touch ${bam_file.baseName}.vcf.gz.tbi
+        fi
+
         """
 }
